Update README.md
Everyone is familiar with the way jquery promise object works. And the use jquery 'deferred', 'then', 'when', 'done', 'fail', 'finally'.

If I have three async functions a,b,c and if i want the following execution flow

1. call a and b initially.
2. on success of both a and b call c.

I will use jquery promise object with 'when'.

But let us take a case with 8 async functions:
a,b,c,d,e,f,g and h.
and if I want to have the following execution flow
1. Call a and d.
2. On success of a, call b.
3. On success of d , call c and e.
4. On error of b, call f.
5. ON success of c and b, call g.
6. on success of b and e call h.
7. And after all responses are received and when there is no open request do SOMETHING.

It becomes difficult to achieve the same using jquery promise.
And as the complexity of the sequence increases, the readability of the code decreases significantly.

Hence, asynchCallsManager.js which helps you seperate the function definitions from the execution flow. 
It helps you define the sequence in which the functions should be called and then manages the execution flow.